slight restructure
