+++ storage/filesystem/object.go
@@ -402,7 +402,7 @@ func (s *ObjectStorage) getFromUnpacked(h plumbing.Hash) (obj plumbing.EncodedOb
 	}
 
 	if size > packfile.LargeObjectThreshold {
-		obj = objfile.NewEncodedObject(s.dir, h, t, size)
+		obj = dotgit.NewEncodedObject(s.dir, h, t, size)
 		s.objectCache.Put(obj)
 		return obj, nil
 	}
